#60 refactor
